# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Rental Car POS & Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles/{vehicleId}/maintenance:
    get:
      operationId: listMaintenanceLogs
      tags: [maintenance]
      summary: List maintenance logs for a vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceLog"
    post:
      operationId: createMaintenanceLog
      tags: [maintenance]
      summary: Create a maintenance log entry
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaintenanceLogBody"
      responses:
        "201":
          description: Created log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceLog"

  # ─── CUSTOMERS ────────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ─── BOOKINGS ─────────────────────────────────────────────────────────────────
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, active, completed, cancelled, null]
        - name: vehicleId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/checkin:
    post:
      operationId: checkinBooking
      tags: [bookings]
      summary: Check-in a booking (record start KM)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinBody"
      responses:
        "200":
          description: Updated booking after check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/checkout:
    post:
      operationId: checkoutBooking
      tags: [bookings]
      summary: Check-out a booking (record end KM + additional charges)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutBody"
      responses:
        "200":
          description: Updated booking after checkout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ─── TRANSACTIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MaintenanceLog:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        type:
          type: string
          enum: [oil_change, tax_renewal, periodic_service, repair, other]
        description:
          type: string
        cost:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - vehicleId
        - type
        - description
        - cost
        - dueDate
        - completedAt
        - createdAt
    CreateMaintenanceLogBody:
      type: object
      properties:
        type:
          type: string
          enum: [oil_change, tax_renewal, periodic_service, repair, other]
        description:
          type: string
        cost:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
      required:
        - type
        - description
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        idNumber:
          type: string
        idType:
          type: string
          enum: [ktp, sim, passport]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        totalBookings:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - phone
        - email
        - idNumber
        - idType
        - address
        - notes
        - totalBookings
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        idNumber:
          type: string
        idType:
          type: string
          enum: [ktp, sim, passport]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name
        - phone
        - idNumber
        - idType
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        idNumber:
          type: string
        idType:
          type: string
          enum: [ktp, sim, passport]
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    Booking:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        vehicleId:
          type: integer
        customerName:
          type: string
        vehicleName:
          type: string
        vehiclePlate:
          type: string
        rentalType:
          type: string
          enum: [self_drive, with_driver]
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        startDate:
          type: string
        endDate:
          type: string
        startKm:
          type: ["integer", "null"]
        endKm:
          type: ["integer", "null"]
        baseAmount:
          type: number
        lateFee:
          type: ["number", "null"]
        washFee:
          type: ["number", "null"]
        damageFee:
          type: ["number", "null"]
        otherFee:
          type: ["number", "null"]
        pickupDropoffFee:
          type: ["number", "null"]
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerId
        - vehicleId
        - customerName
        - vehicleName
        - vehiclePlate
        - rentalType
        - status
        - startDate
        - endDate
        - startKm
        - endKm
        - baseAmount
        - lateFee
        - washFee
        - damageFee
        - otherFee
        - pickupDropoffFee
        - totalAmount
        - notes
        - createdAt
    CreateBookingBody:
      type: object
      properties:
        customerId:
          type: integer
        vehicleId:
          type: integer
        rentalType:
          type: string
          enum: [self_drive, with_driver]
        startDate:
          type: string
        endDate:
          type: string
        notes:
          type: ["string", "null"]
        pickupDropoffFee:
          type: ["number", "null"]
      required:
        - customerId
        - vehicleId
        - rentalType
        - startDate
        - endDate
    UpdateBookingBody:
      type: object
      properties:
        rentalType:
          type: string
          enum: [self_drive, with_driver]
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        pickupDropoffFee:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    CheckinBody:
      type: object
      properties:
        startKm:
          type: integer
      required:
        - startKm
    CheckoutBody:
      type: object
      properties:
        endKm:
          type: integer
        lateFee:
          type: ["number", "null"]
        washFee:
          type: ["number", "null"]
        damageFee:
          type: ["number", "null"]
        otherFee:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - endKm
...